pyproject.toml
[tool.poetry]
name = "chat-langchain"
version = "0.1.0"
description = ""
authors = ["SN <6432132+samnoyes@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.28.0"
fastapi = "^0.103.1"
pydantic = "1.10"
langchain = "^0.0.333"
uvicorn = "^0.23.2"
beautifulsoup4 = "^4.12.2"
tiktoken = "^0.4.0"
weaviate-client = "^3.23.2"
psycopg2 = "^2.9.7"
lxml = "^4.9.3"
langserve = {extras = ["server"], version = "^0.0.34"}
voyageai = "^0.1.4"
python-dotenv = "^1.0.0"
langchain-community = "^0.0.3"
chroma = "^0.2.0"
chromadb = "^0.4.20"
torch = "^2.1.2"
transformers = "^4.36.1"


[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
